Rework application model to use correct API of the application manager
The application model now gets automatically updated once a launch point within the
application manager is removed, added or changed. Furthermore this includes now handling
of all different launch point types which can be also a link to a website. Applications
which should be hidden from the launcher like the first use application are now correctly
hidden as they are not having a launch point.

Signed-off-by: Simon Busch <morphis@gravedo.de>